benchmarks/framework-shootout/apps/tornado_app/requirements.txt
tornado==6.4.*
psycopg2-binary==2.9.*
orjson==3.*
